cleanup startup Split out the response
