+++ src/main/java/io/mapsmessaging/rest/api/impl/discovery/DiscoveryManagementApi.java
@@ -1,24 +1,36 @@
+/*
+ * Copyright [ 2020 - 2024 ] [Matthew Buckton]
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
 package io.mapsmessaging.rest.api.impl.discovery;
 
 import io.mapsmessaging.MessageDaemon;
-import io.mapsmessaging.engine.destination.DestinationImpl;
-import io.mapsmessaging.network.discovery.AdapterManager;
 import io.mapsmessaging.rest.api.impl.BaseRestApi;
-import io.mapsmessaging.rest.data.destination.Destination;
-import io.mapsmessaging.rest.responses.DestinationResponse;
+import io.mapsmessaging.rest.data.discovery.DiscoveredServers;
+import io.mapsmessaging.rest.data.discovery.ServiceData;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.ws.rs.GET;
 import jakarta.ws.rs.Path;
 import jakarta.ws.rs.Produces;
 import jakarta.ws.rs.core.MediaType;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 import javax.jmdns.ServiceInfo;
-
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Map;
 
 import static io.mapsmessaging.rest.api.Constants.URI_PATH;
 
@@ -70,40 +82,5 @@ public class DiscoveryManagementApi extends BaseRestApi {
       discoveredServers.add(new DiscoveredServers(entry.getKey(), discoveredServiceData));
     }
     return discoveredServers;
-  }
-
-  @Getter
-  @Setter
-  @NoArgsConstructor
-  @AllArgsConstructor
-  public static final class DiscoveredServers {
-    private String server;
-    private List<ServiceData> serviceInfo;
-  }
-
-  @Getter
-  @Setter
-  public static final class ServiceData {
-
-    private String application;
-    private String name;
-    private String domain;
-    private String[] hostAddresses;
-    private int port;
-    private Map<String, String> properties;
-
-    public ServiceData(ServiceInfo serviceInfo) {
-      hostAddresses = serviceInfo.getHostAddresses();
-      port = serviceInfo.getPort();
-      name = serviceInfo.getName();
-      application = serviceInfo.getApplication();
-      domain = serviceInfo.getDomain();
-      properties = new LinkedHashMap<>();
-      Enumeration<String> propertyNames = serviceInfo.getPropertyNames();
-      while (propertyNames.hasMoreElements()) {
-        String name = propertyNames.nextElement();
-        properties.put(name, serviceInfo.getPropertyString(name));
-      }
-    }
   }
 }
