+++ src/main/java/io/mapsmessaging/rest/data/discovery/ServiceData.java
@@ -1,0 +1,53 @@
+/*
+ * Copyright [ 2020 - 2024 ] [Matthew Buckton]
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+
+package io.mapsmessaging.rest.data.discovery;
+
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.jmdns.ServiceInfo;
+import java.util.Enumeration;
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+@Getter
+@Setter
+public class ServiceData {
+
+  private String application;
+  private String name;
+  private String domain;
+  private String[] hostAddresses;
+  private int port;
+  private Map<String, String> properties;
+
+  public ServiceData(ServiceInfo serviceInfo) {
+    hostAddresses = serviceInfo.getHostAddresses();
+    port = serviceInfo.getPort();
+    name = serviceInfo.getName();
+    application = serviceInfo.getApplication();
+    domain = serviceInfo.getDomain();
+    properties = new LinkedHashMap<>();
+    Enumeration<String> propertyNames = serviceInfo.getPropertyNames();
+    while (propertyNames.hasMoreElements()) {
+      String key = propertyNames.nextElement();
+      properties.put(key, serviceInfo.getPropertyString(key));
+    }
+  }
+}
